Issue #124 -- Register for an Event
Updated the table so that it displays a participants age where the music level used to be displayed. Added a new column labeled 'Music Level' with drop-down selections for music levels. The music level stays highlighted for all test values (1-12) and highlighting is removed when a user selects the placeholder value. This page needs further work since changes to the music level affect all rows.