Issue # 157 -- Update CSS Forms and Layout
Updated many aspects of the non-portal pages CSS so that the forms are easier to read. Input fields and buttons were updated. Additionally, the forms were all changed so that they are centered on the page and the negative width was updated to fix the footer so that it is shown properly on the page.